terraform/functions/login/requirements.txt
boto3==1.35.76
bcrypt==4.2.1
PyJWT==2.10.1
